reverse_search.py
"""
Reverse Search Module
---------------------
Performs fingerprint-based internet lookup.
"""

import imagehash
from PIL import Image
import requests

def reverse_image_search(image_path: str) -> dict:
    """
    Placeholder logic for reverse image search.
    In real use: Google Lens, Bing API, or open datasets.
    """
    try:
        img = Image.open(image_path)
        phash = imagehash.phash(img)

        # Placeholder response (simulate internet signal)
        return {
            "found": True,
            "match_confidence": 0.65,
            "sources": ["unknown_blog", "ai_art_platform"]
        }
    except Exception:
        return {
            "found": False,
            "match_confidence": 0.0,
            "sources": []
        }
