requirements.txt
# ===============================
# Core Language & Utilities
# ===============================
numpy>=1.24.0
scipy>=1.10.0
requests>=2.31.0
tqdm>=4.66.0

# ===============================
# Machine Learning (Core)
# ===============================
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
scikit-learn>=1.3.0

# ===============================
# Audio Processing
# ===============================
librosa>=0.10.1
soundfile>=0.12.1
pydub>=0.25.1

# ===============================
# Image Processing
# ===============================
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.22.0
imagehash>=4.3.1

# ===============================
# Video Processing
# ===============================
ffmpeg-python>=0.2.0
moviepy>=1.0.3

# ===============================
# Face & Landmark Analysis
# ===============================
mediapipe>=0.10.8
dlib>=19.24.0

# ===============================
# Metadata & Forensics
# ===============================
exifread>=3.0.0
hachoir>=3.2.0

# ===============================
# Internet Verification
# ===============================
beautifulsoup4>=4.12.2
aiohttp>=3.9.0
tldextract>=5.1.0

# ===============================
# API / Interface (Optional Phase)
# ===============================
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6.0

# ===============================
# Logging & Testing
# ===============================
pytest>=8.0.0
black>=24.1.0
flake8>=7.0.0
